@@ -3,35 +3,34 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "axios": "^0.21.1",
-    "dotenv": "^9.0.2",
-    "google-maps-react": "^2.0.6",
-    "jwt-decode": "^3.1.2",
-    "material-ui-image": "^3.3.1",
-    "node-ipinfo": "^2.0.1",
-    "notistack": "^1.0.3",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/icons-material": "^6.4.5",
+    "@mui/material": "^6.4.5",
+    "@mui/styles": "^6.4.5",
+    "@react-google-maps/api": "^2.20.6",
+    "@reduxjs/toolkit": "^2.5.1",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.2.0",
+    "@testing-library/user-event": "^14.6.1",
+    "axios": "^1.7.9",
+    "jwt-decode": "^4.0.0",
+    "notistack": "^3.0.2",
+    "os-browserify": "^0.3.0",
+    "path-browserify": "^1.0.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-geocode": "^0.2.2",
-    "react-geolocated": "^3.2.0",
-    "react-geolocation": "^1.0.4",
-    "react-google-charts": "^3.0.15",
-    "react-minimal-pie-chart": "^8.1.0",
-    "react-redux": "^7.2.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "recharts": "^2.0.8",
-    "redux": "^4.0.5",
+    "react-google-charts": "^5.2.1",
+    "react-minimal-pie-chart": "^9.1.0",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^6.29.0",
+    "react-scripts": "^5.0.1",
+    "recharts": "^2.15.1",
+    "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0",
-    "web-vitals": "^0.2.4"
+    "redux-thunk": "^3.1.0",
+    "web-vitals": "^4.2.4"
   },
   "scripts": {
     "start": "react-scripts start",
