@@ -1,7 +1,7 @@
 import React,{useEffect,useState} from "react";
 import {useDispatch,useSelector} from "react-redux";
-import {Typography,Grid,TextField,makeStyles,Paper,Divider} from "@material-ui/core";
-import {Map,GoogleApiWrapper} from 'google-maps-react';
+import {Typography,Grid,TextField,Paper,Divider} from "@mui/material";
+import { makeStyles } from '@mui/styles';
 import Geocode from "react-geocode";
 
 const useStyles = makeStyles(theme=>({
