@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from "react-dom/client";
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
@@ -7,15 +7,16 @@ import {SnackbarProvider} from 'notistack';
 import {BrowserRouter} from "react-router-dom";
 import {ThemeContextProvider} from "./context/ThemeContext";
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+root.render(
   <ThemeContextProvider>
       <SnackbarProvider maxSnack={1}>
           <BrowserRouter>
               <App/>
           </BrowserRouter>
       </SnackbarProvider>
-  </ThemeContextProvider>,
-  document.getElementById('root')
+  </ThemeContextProvider>
 );
 
 // If you want to start measuring performance in your app, pass a function
